Clean up viewManager, move certain functions into it
