+++ src/main/app/app.test.js
@@ -6,7 +6,6 @@ import {app, dialog} from 'electron';
 import CertificateStore from 'main/certificateStore';
 import ViewManager from 'main/views/viewManager';
 import MainWindow from 'main/windows/mainWindow';
-import WindowManager from 'main/windows/windowManager';
 
 import {handleAppWillFinishLaunching, handleAppCertificateError, certificateErrorCallbacks} from 'main/app/app';
 import {getDeeplinkingURL, openDeepLink} from 'main/app/utils';
@@ -20,13 +19,6 @@ jest.mock('electron', () => ({
     dialog: {
         showMessageBox: jest.fn(),
     },
-}));
-
-jest.mock('common/servers/serverManager', () => ({
-    getServer: (name) => ({
-        name,
-        url: 'http://server-1.com',
-    }),
 }));
 
 jest.mock('main/app/utils', () => ({
@@ -55,6 +47,7 @@ jest.mock('main/windows/mainWindow', () => ({
 }));
 jest.mock('main/views/viewManager', () => ({
     getView: jest.fn(),
+    getViewByWebContentsId: jest.fn(),
 }));
 
 describe('main/app/app', () => {
@@ -104,10 +97,19 @@ describe('main/app/app', () => {
         const mainWindow = {};
         const promise = Promise.resolve({});
         const certificate = {};
+        const view = {
+            tab: {
+                server: {
+                    name: 'test-team',
+                    url: new URL(testURL),
+                },
+            },
+            load: jest.fn(),
+        };
 
         beforeEach(() => {
             MainWindow.get.mockReturnValue(mainWindow);
-            WindowManager.getServerNameByWebContentsId.mockReturnValue('test-team');
+            ViewManager.getViewByWebContentsId.mockReturnValue(view);
         });
 
         afterEach(() => {
@@ -166,12 +168,9 @@ describe('main/app/app', () => {
 
         it('should load URL using MattermostView when trusting certificate', async () => {
             dialog.showMessageBox.mockResolvedValue({response: 0});
-            const load = jest.fn();
-            ViewManager.getView.mockReturnValue({load});
-            WindowManager.getViewIdByWebContentsId.mockReturnValue('view-name');
             await handleAppCertificateError(event, webContents, testURL, 'error-1', certificate, callback);
             expect(callback).toHaveBeenCalledWith(true);
-            expect(load).toHaveBeenCalledWith(testURL);
+            expect(view.load).toHaveBeenCalledWith(testURL);
         });
 
         it('should explicitly untrust if user selects More Details and then cancel with the checkbox checked', async () => {
