+++ src/main/app/utils.ts
@@ -14,6 +14,7 @@ import {Boundaries} from 'types/utils';
 
 import Config from 'common/config';
 import JsonFileManager from 'common/JsonFileManager';
+import ServerManager from 'common/servers/serverManager';
 import {MattermostServer} from 'common/servers/MattermostServer';
 import urlUtils from 'common/utils/url';
 import {APP_MENU_WILL_CLOSE} from 'common/communication';
@@ -24,8 +25,8 @@ import {localizeMessage} from 'main/i18nManager';
 import {createMenu as createAppMenu} from 'main/menus/app';
 import {createMenu as createTrayMenu} from 'main/menus/tray';
 import {ServerInfo} from 'main/server/serverInfo';
-import ServerManager from 'common/servers/serverManager';
 import {setTrayMenu} from 'main/tray/tray';
+import ViewManager from 'main/views/viewManager';
 import WindowManager from 'main/windows/windowManager';
 import MainWindow from 'main/windows/mainWindow';
 
@@ -55,7 +56,7 @@ export function handleUpdateMenuEvent() {
     const aMenu = createAppMenu(Config, updateManager);
     Menu.setApplicationMenu(aMenu);
     aMenu.addListener('menu-will-close', () => {
-        WindowManager.focusBrowserView();
+        ViewManager.focusCurrentView();
         WindowManager.sendToRenderer(APP_MENU_WILL_CLOSE);
     });
