+++ src/main/downloadsManager.test.js
@@ -80,6 +80,7 @@ jest.mock('main/notifications', () => ({}));
 jest.mock('main/windows/windowManager', () => ({
     sendToRenderer: jest.fn(),
 }));
+jest.mock('main/views/viewManager', () => ({}));
 jest.mock('common/config', () => {
     const original = jest.requireActual('common/config');
     return {
