+++ src/main/views/modalManager.ts
@@ -22,6 +22,7 @@ import {
 
 import {getAdjustedWindowBoundaries} from 'main/utils';
 import WebContentsEventManager from 'main/views/webContentEvents';
+import ViewManager from 'main/views/viewManager';
 import WindowManager from 'main/windows/windowManager';
 
 import {ModalView} from './modalView';
@@ -113,7 +114,7 @@ export class ModalManager {
             this.showModal();
         } else {
             WindowManager.sendToRenderer(MODAL_CLOSE);
-            WindowManager.focusBrowserView();
+            ViewManager.focusCurrentView();
         }
     }
